docker_image/requirements.txt
prometheus_client
requests
httpx
pyyaml